feat(portfolio, cart): Implement SIP Step-Up, top-up for existing funds, and enhanced trading account flow
This commit introduces major enhancements to the mutual fund investment journey, specifically focusing on the ability to "Step Up" existing SIPs, top up existing folio funds, and a more robust trading account setup process with prefilled user data. It also adds several "About" informational screens for the Velvet platform and FIRE reports.

**Key Changes:**

*   **Mutual Fund Portfolio & Folio Management:**
    *   **New Folio Screens:** Introduced `ExistingFundScreen` and `FolioFundMFScreen` to view funds filtered by folio.
    *   **Top-Up Flow:** Added `ExistingFundLumpSumScreen` allowing users to select multiple existing funds and add lump sum amounts (Top-up) in a single transaction.
    *   **Folio Fund UI:** Refactored `MutualFundsCard` into `FolioFundCard` with a cleaner layout focusing on current value and folio details.

*   **Shopping Cart & SIP Step-Up:**
    *   **SIP Step-Up:** Implemented a "Step Up" feature in the cart, allowing users to automatically increase their SIP amount by a percentage or fixed amount. Includes a percentage slider and minimum amount validation.
    *   **Cart Refactoring:** Moved cart logic to a dedicated `cart` package. Split the cart view into `LumpSumCartContent` and `SIPCartContent`.
    *   **Data Models:** Updated `SipItemDomain` to support step-up parameters (`stepUpAmount`, `minStepUpPercent`, etc.).

*   **Trading Account & KYC:**
    *   **Prefilled Data:** Integrated `GetTradingAccountPrefilledDataUseCase` to automatically populate the trading account form with user profile data (Name, PAN, DOB, Address).
    *   **Confirmation Flow:** Updated `tradingAccountConfirmation` to accept mandatory regulatory fields like tax status, holding nature, and guardian details for minors.
    *   **Dynamic Logic:** Automatically maps Source of Wealth to specific Occupation Types for regulatory compliance.

*   **Profile & Content:**
    *   **Informational Screens:** Added comprehensive `AboutVelvetScreen` and `AboutFireScreen` providing detailed insights into the platform's philosophy and financial independence reports.
    *   **Insurance UI:** Enhanced `TermInsurance` and `HealthInsurance` screens with `HorizontalPager` to display multiple educational banners.

*   **Technical Refinements:**
    *   **API Integration:** Added `investMoreLumpsum` and `getFolioFunds` to `UserFinanceRepo`.
    *   **Architecture:** Migrated several ViewModels to use Koin's `viewModel` factory and added necessary UseCases for folio management and cart operations.
    *   **UI Components:** Added `BlostemFooter` to Fixed Deposit screens to indicate the service provider. Refined `WorthCard` visibility toggle logic.
    *   **Fixes:** Updated `GoalMapper` to handle scheme IDs as strings for API compatibility and fixed keyboard padding issues in footer components.

+++ composeApp/src/commonMain/kotlin/org/sharad/velvetinvestment/presentation/goals/compose/MapSchemesScreen.kt
@@ -395,7 +395,7 @@ fun MapSchemesBottomSheetContent(
                             items(data) { scheme ->
                                 SelectableSchemeItem(
                                     scheme = scheme,
-                                    onToggle = { viewModel.toggleSelection(scheme.schemeId.toInt()) }
+                                    onToggle = { viewModel.toggleSelection(scheme.schemeId) }
                                 )
                             }
                         }
